Rename internal/utils to internal/osutil
