+++ internal/osutil/doc.go
@@ -1,0 +1,2 @@
+// Package osutil provides some OS-level helper functions.
+package osutil
