--- internal/utils/doc.go
@@ -1,4 +1,0 @@
-// Package utils provides some file- and filepath-helper functions.
-//
-// TODO: `utils` is too vague. Find a better name for this package.
-package utils
